+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### Changed
+- The agent now supports reporting the time for launches, test items and logs with microsecond precision in the ISO string format.
+For logs, microsecond precision is available on the UI from ReportPortal version 24.2.
+- `@reportportal/client-javascript` bumped to version `5.3.0`.
 
 ## [5.1.10] - 2024-09-17
 ### Changed